requirements.txt
argon2-cffi>=18.3.0
bcrypt>=3.1.4
Django>=2.1.3
djangorestframework>=3.9.0
drf-enum-field>=0.9.2